putMarker

+++ app/src/main/java/kr/co/lion/android01/mapmemoproject/UserInfoAll.kt
@@ -6,6 +6,6 @@ data class UserInfoAll (
     var date:String,
     var title:String,
     var contents:String,
-    var latitude:Float,
-    var longitude:Float
+    var latitude:Double,
+    var longitude:Double
 )
